+++ storage_schema.sql
@@ -93,7 +93,7 @@ CREATE TABLE `st_files` (
   KEY `fk_last_scan` (`last_scan_id`),
   CONSTRAINT `fk_last_scan` FOREIGN KEY (`last_scan_id`) REFERENCES `st_scans` (`scan_id`) ON DELETE SET NULL,
   CONSTRAINT `st_files_drive_id_fk` FOREIGN KEY (`drive_id`) REFERENCES `st_drives` (`id`) ON DELETE CASCADE ON UPDATE CASCADE
-) ENGINE=InnoDB AUTO_INCREMENT=1730471 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
+) ENGINE=InnoDB AUTO_INCREMENT=1742150 DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_general_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
 
 --
@@ -114,6 +114,9 @@ CREATE TABLE `st_recovery` (
   `output_data` longblob DEFAULT NULL,
   `output_text` longtext DEFAULT NULL,
   `description` text DEFAULT NULL,
+  `success` tinyint(1) DEFAULT NULL,
+  `error_message` text DEFAULT NULL,
+  `return_code` int(11) DEFAULT NULL,
   PRIMARY KEY (`recovery_id`),
   KEY `idx_drive_id` (`drive_id`),
   KEY `idx_scan_id` (`scan_id`),
@@ -194,4 +197,4 @@ CREATE TABLE `st_version` (
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2025-09-10  1:43:55
+-- Dump completed on 2025-09-10  2:55:30